refactored fish-movement-input-handling lines of code from GameStageState to Fish class. Changed Entity, Creature, Fish, and OverworldCursor class to use x and y coordinate values as float instead of int (same with xMove, yMove, and speed). This change affected SaverAndLoader's calls to writeInt() and readInt(), which are now calls to writeFloat() and readFloat(). The Creature class now has int health and float speed member fields, and public constants for DEFAULT_HEALTH and DEFAULT_SPEED. The Fish class's getInput() method (where fish-movement-input-handling lines of code from GameStageState moved to) utilizes the speed variable to set its xMove and yMove based on KeyManager's directional boolean.

+++ src/com/evo/entities/moveable/OverWorldCursor.java
@@ -8,7 +8,7 @@ import java.awt.image.BufferedImage;
 
 public class OverWorldCursor extends Creature {
 
-    public OverWorldCursor(Handler handler, BufferedImage image, int x, int y) {
+    public OverWorldCursor(Handler handler, BufferedImage image, float x, float y) {
         super(handler, image, x, y, image.getWidth(), image.getHeight());
     }
 
@@ -17,7 +17,7 @@ public class OverWorldCursor extends Creature {
     }
 
     public void render(Graphics g) {
-        g.drawImage(image, x, y, width, height,null);
+        g.drawImage(image, (int)x, (int)y, width, height,null);
     }
 
 }
